add image_process class
