+++ cal.h
@@ -4,6 +4,8 @@
 #include <vector>
 #include <sstream>
 #include <ctime>
+#include <tesseract/baseapi.h>
+#include <allheaders.h>
 #pragma once
 
 namespace ical {
@@ -11,7 +13,9 @@ namespace ical {
 class calendar {
     public:
         std::string timezone, DTSTART, DTEND, freq, weekstop, untillD, byday, location, Ename;
-
+        
+        // Default constructor
+        calendar();
         // Constructor
         calendar(std::string TZ, std::string WT, std::string f) {
             timezone = TZ;
@@ -110,6 +114,7 @@ class calendar {
             if (s == "WE") return "WE";
             if (s == "SA") return "SA";
             if (s == "SU") return "SU";
+            if (s == "TU") return "TU";
             return "";       
         }
 
@@ -139,6 +144,63 @@ class calendar {
             }
             return sArr; 
         }
+};
+
+// all this shit could be cleaner and better.
+class image_process {
+    public:
+        // Function to convert char* to vector<string> by splitting char* on '\\n'
+        std::vector<std::string> tostrVector(char * in) {
+            std::vector<std::string> text;
+            size_t i = 0, j = 0;
+
+            while (in[i] != '\0') {
+                if (in[i] == '\n') {
+                    size_t line_size = &in[i] - &in[j];
+                    std::string t = "";
+                    // this work but there are overflow in this j-1 when j = 0, maybe need a better way to write this.
+                    for (size_t a = j; a < line_size + (j-1); a++) t.push_back(in[a]);
+                    text.push_back(t);
+                    j = i+1;
+                }
+                i++;
+            }
+
+            return text;
+        }
+
+        // Function to process img and return vector of string.
+        std::vector<std::string> process_Image(const char* imgPath) {
+            char * outText;
+            Pix *image  = pixRead(imgPath);
+            std::vector<std::string> out;
+
+            tesseract::TessBaseAPI *api = new tesseract::TessBaseAPI();
+            if (api->Init("/Projects/cpp/CPEPROJECT/Uics/tessdata", "eng")) {
+                std::cout << "Could not initialize tesseract.\n";
+                exit(1);
+            }
+
+            api->SetImage(image);
+            outText = api->GetUTF8Text();
+            out = tostrVector(outText);
+
+            delete [] outText;
+            api->End();
+            delete api;
+            pixDestroy(&image);
+
+            return out;
+        }
+
+        // simple func to create event using vector of process_image()
+        void img_events(std::ofstream &file, const char* imgPath, calendar obj) {
+            std::vector<std::string> in = process_Image(imgPath);
+            for (size_t i = 0; i < in.size(); i++) {   
+                obj.createEvent(file, in[i], 1);
+            }
+        }
+
 };
 
 } // end namespace ical
